+++ js/data.js
@@ -1,0 +1,82 @@
+import {
+  getRandomInteger,
+  getRandomFloatWithPrecision,
+  getRandomValueFromArray,
+  getRandomLengthArray,
+} from './util.js';
+
+const TITLE = 'Объявление об аренде недвижимости';
+const DESCRIPTION = 'Самое крутое предложение';
+const MIN_PRICE = 1;
+const MAX_PRICE = 1000000;
+const MIN_AVATAR_NUMBER = 1;
+const MAX_AVATAR_NUMBER = 8;
+const MIN_ROOMS_NUMBER = 1;
+const MAX_ROOMS_NUMBER = 10;
+const MIN_GUESTS_NUMBER = 1;
+const MAX_GUESTS_NUMBER = 20;
+const MIN_LATITUDE = 35.65000;
+const MAX_LATITUDE = 35.70000;
+const MIN_LONGITUDE = 139.70000;
+const MAX_LONGITUDE = 139.80000;
+const LOCATION_PRECISION = 5;
+
+
+const CHECKINTIMES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+function getAvatarUrl() {
+  const number = getRandomInteger(MIN_AVATAR_NUMBER , MAX_AVATAR_NUMBER);
+  return `img/avatars/user0${number}.png`;
+}
+
+function getOffer() {
+  const ad = {
+    author: {
+      avatar: getAvatarUrl(),
+    },
+    location: {
+      x: getRandomFloatWithPrecision(MIN_LATITUDE, MAX_LATITUDE, LOCATION_PRECISION),
+      y: getRandomFloatWithPrecision(MIN_LONGITUDE, MAX_LONGITUDE, LOCATION_PRECISION),
+    },
+    offer: {
+      title: TITLE,
+      price: getRandomInteger(MIN_PRICE, MAX_PRICE),
+      type: getRandomValueFromArray(TYPES),
+      rooms: getRandomInteger(MIN_ROOMS_NUMBER, MAX_ROOMS_NUMBER),
+      guests: getRandomInteger(MIN_GUESTS_NUMBER, MAX_GUESTS_NUMBER),
+      checkin: getRandomValueFromArray(CHECKINTIMES),
+      checkout: getRandomValueFromArray(CHECKINTIMES),
+      features: getRandomLengthArray(FEATURES),
+      description: DESCRIPTION,
+      photos: getRandomLengthArray(PHOTOS),
+    },
+  }
+
+  ad.offer.address = `${ad.location.x}, ${ad.location.y}`;
+  return ad;
+}
+
+export { getOffer }
